+++ src/state/store.ts
@@ -10,6 +10,7 @@ import {
   createAppearanceSlice,
 } from './slices/appearanceSlice'
 import { AUTH_INITIAL, type AuthSlice, createAuthSlice } from './slices/authSlice'
+import { CALLOUTS_INITIAL, type CalloutsSlice, createCalloutsSlice } from './slices/calloutsSlice'
 import { CAMERA_INITIAL, type CameraSlice, createCameraSlice } from './slices/cameraSlice'
 import {
   CAMERA_VIEWS_INITIAL,
@@ -162,6 +163,7 @@ export interface RootState
     UserSetsSlice,
     UserStylesSlice,
     RecentSlice,
+    CalloutsSlice,
     FavouritesSlice,
     SavedMaterialsSlice,
     ColorPaletteSlice,
@@ -203,6 +205,7 @@ const INITIAL = {
   ...AUTH_INITIAL,
   ...USER_STYLES_INITIAL,
   ...RECENT_INITIAL,
+  ...CALLOUTS_INITIAL,
   ...FAVOURITES_INITIAL,
   ...SAVED_MATERIALS_INITIAL,
   ...COLOR_PALETTE_INITIAL,
@@ -245,6 +248,7 @@ export const useStore = create<RootState>((set, get, api) => ({
   ...createFeatureFlagsSlice(set, get, api),
   ...createAuthSlice(set, get, api),
   ...createRecentSlice(set, get, api),
+  ...createCalloutsSlice(set, get, api),
   ...createFavouritesSlice(set, get, api),
   ...createSavedMaterialsSlice(set, get, api),
   ...createColorPaletteSlice(set, get, api),